hotfix 2.0.3 forgot to set the global selected operator
