+++ Ladefuchs/app/build.gradle
@@ -11,8 +11,8 @@ android {
         applicationId "app.ladefuchs.android"
         minSdkVersion 26
         targetSdkVersion 33
-        versionCode 202
-        versionName "2.0.2"
+        versionCode 203
+        versionName "2.0.3"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
