+++ Ladefuchs/app/src/main/res/layout/fragment_about.xml
@@ -5,8 +5,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@drawable/rounded_header_card"
-    tools:context=".ui.about.AboutFragment">
+    android:background="@drawable/rounded_header_card">
 
     <ImageButton
         android:id="@+id/back_button"
@@ -668,8 +667,8 @@
                 android:id="@+id/handcraftedText"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-
                 android:layout_marginTop="3dp"
+                android:layout_marginEnd="5dp"
                 android:fontFamily="@font/bitter"
                 android:layout_marginBottom="20dp"
                 android:text="@string/handcrafted"
