+++ Ladefuchs/app/src/main/res/values/strings.xml
@@ -68,7 +68,7 @@
     <string name="mastodon_malik"><a href="https://mastodon.social/@Malik">@Malik@mastodon.social</a></string>
     <string name="twitter_thorsten"><a href="https://twitter.com/Thorsten_1">@Thorsten_1</a></string>
     <string name="twitter_roddi"><a href="https://twitter.com/roddi">@roddi</a></string>
-    <string name="handcrafted">Handgefertigt aus ♥️ zur Elektromobilität in \n👑 Aachen, 🥨 Fürstenfeldbruck,\n🏰 Ludwigsburg, ⚒️ Ahlen und 🐻 Berlin.</string>
+    <string name="handcrafted">Handgefertigt aus ♥️ zur Elektromobilität in 👑 Aachen, 🥨 Fürstenfeldbruck, 🏰 Ludwigsburg, ⚒️ Ahlen und 🐻 Berlin.</string>
     <string name="imprint_header">Impressum</string>
     <string name="acknowledgement_header">DrittLizenzen</string>
     <string name="imprint_text">Angaben gemäß § 5 TMG:\n\nBastian Wölfle\nSchwabenstraße 11\n82256 Fürstenfeldbruck\n\nKontakt:\info@schlingel.de\n\nVerantwortlich für den Inhalt nach § 55 Abs. 2 RStV:\nBastian Wölfle\nSchwabenstraße 11\n82256 Fürstenfeldbruck</string>
